+++ packages/cca-scripts/scripts/get-client-config.js
@@ -1,8 +1,9 @@
 const util = require('./util');
 
 util.cloudformation.getStackOutputs().then(outputs => {
-	const config = {
-		baseUrl: outputs.ApiUrl
+	let config = {
+		baseUrl: outputs.apiUrl,
+		apiKey: outputs.apiKey
 	};
 
 	console.log(JSON.stringify(config, null, 2));