+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## December 28, 2017
 
+#### `cca-scripts@1.2.4`
+
+* Add `apiKey` to `get-client-config` output
+
+## December 28, 2017
+
 #### `create-cloud-api@1.0.3`
 
 * Added CHANGELOG.md